Release the python REST client for REST version 1.7 This accompanies the PurityFB//2.3.4 release.
This release adds new options to the directory-services/
endpoint to support configuration of NIS servers and an
NIS domain for the NFS directory service.

This release also adds support for specifying a CA certificate(s)
file path when instantiating the python REST client, allowing SSL
verification to be enabled at creation time. It also adds support
for changing the path to the CA certificate(s) file at a later time,
and changing it while enabling SSL verification.

New examples have been added for both of the newly supported options.