REST SDK fixes for Release 1.9.2
Several fixes have been made to the client, and are visible in the new Release
Notes section of the README file. Additionally, the Audit functionality which
was introduced in PurityFB//3.0 has now been added to the purity_fb Python SDK.

This commit also includes fixes to documentation and sample code which was
previously incorrect.