This commit contains changes for releasing v1.10.0 of the Python REST SDK.
In addition to the fixes and new features that are mentioned in README.md, this
commit also contains documentation fixes which remove several invalid references
and links (such as those to 'str.md'). We've also begun using a slightly updated
template for generating the Python for models. These changes have, somewhat
unfortunately, contributed to another rather large commit which touches past
versions of the SDK, but should not have any functional effect.